+++ lib/viewNotifier.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/widgets.dart';
 import 'package:udf/stateProvider.dart';
 
+import 'model.dart';
+
 class ViewNotifier<T extends StateProvider> extends InheritedNotifier<StateProvider> {
   ViewNotifier(
     T stateProvider,
     Widget child,
   ) : super(notifier: stateProvider, child: child);
 
-  static T of<T extends StateProvider>(BuildContext context) =>
-      context.dependOnInheritedWidgetOfExactType<ViewNotifier<T>>()?.notifier as T;
+  static M model<M extends Model<M>, T extends StateProvider<M>>(BuildContext context) =>
+      context.dependOnInheritedWidgetOfExactType<ViewNotifier<T>>()!.notifier!.model() as M;
 }