+++ CHANGELOG.md
@@ -83,3 +83,10 @@
 ## [0.0.17] - 5-03-2021
 
 * StateProvider, providerOf needs to be more strict with its typing
+
+
+## [0.0.18] - 5-03-2021
+
+* StateProvider, providerOf needs to be more strict with its typing
+* Refactored the VIewNotifier to provide the model.
+* added test for the viewNotifier